Bullet port compiles. Now to build some sort of unit testing procedures. Also there's going to be massive garbage generation.

+++ Voxelarium.Bullet/LinearMath/Porting/GrahamScan2dConvexHull.cs
@@ -29,7 +29,7 @@ namespace Bullet.LinearMath
 		public GrahamVector3( ref btVector3 org, int orgIndex )
 		{
 			v = org;
-			m_orgIndex = orgIndex;
+			m_orgIndex = (short)orgIndex;
 		}
 
 		public struct btAngleCompareFunc : IComparer<GrahamVector3>
